swupdate: update to 2019.04 and enable libubootenv
Update to 2019.04 dropping upstreamed patch.
Also swupdate now has the ability to use the external library
"libubootenv" to access U-Boot environment. As using it drops the
dependency on the board's U-Boot, it is much more convenient and we
switch to use it by default. Default U-Boot default env should be now
provided in /etc/u-boot-initial-env rather than being embedded in the
library.

Signed-off-by: Diego Rondini <diego.rondini@kynetics.com>