rpi-u-boot-scr: add regular system bootscript
We need a regular system bootscript without the saveenv command
and the storage of the U-Boot environment on the uboot.env file
in the first partition of the SD card.
A robust and flexible update mechanism must allow updating the
first partition (usually the boot partition includes kernel image,
device tree blobs and other important binaries), so the U-Boot
environment must be stored in other ways, such as a dedicated
partition or in an unpartitioned storage area.

Signed-off-by: Jacopo Dall'Aglio <jacopo.dallaglio@kynetics.com>