u-boot: generate and deploy environment for RPi
RPi default is to store U-Boot environment in a dedicated file in
the boot partition, but this is not compatible with Update Factory
update mechanism since the environment will be wiped if the boot
partition is updated.
Therefore we will store the environment in an unpartitioned space
in the SD card, at 4 KiB offset from the storage first block.
To do that we need to generate the environment binary from the
u-boot-initial-env fallback environment file already generated by
setting the bitbake variable UBOOT_INITIAL_ENV.

We also need to disable the default config option:
- CONFIG_ENV_FAT_DEVICE_AND_PART
while setting:
- CONFIG_ENV_IS_IN_MMC=y
- CONFIG_SYS_MMC_ENV_DEV=0
- CONFIG_ENV_OFFSET=0x1000
to store the environment in the SD card at 4 KiB offset.

We also provide the fw_config.env file that points to the
unpartitioned storage area where the U-Boot environment is stored.

Signed-off-by: Jacopo Dall'Aglio <jacopo.dallaglio@kynetics.com>