initscripts-swupdate-uf: handle errors reading U-Boot environment
If FW utils are run too early in the init script the storage
where the U-Boot environment is located might be not yet fully
initialized and exposed as a device file.
There's also the unlikely ocurrence that the U-Boot environment
gets corrupted when booting the Recovery OS ramdisk.

For these reasons we added an error control code that tries to
read the U-Boot environment at least 30 times with 100 ms delay
between attempts before giving up and reporting back the error.

Signed-off-by: Jacopo Dall'Aglio <jacopo.dallaglio@kynetics.com>