swu-image.inc: include to create a swu image recipe
This include file must be required in the .swu image recipe.
The sw-decription file must be provided for each image and it
must contain a number of image entries equal to the number of
extensions listed in UF_SWUPDATE_IMAGES_FSTYPES.
UF_SWUPDATE_IMAGES_FSTYPES can be customized and UF_IMAGE must
correspond to the .swu target image.

SWUPDATE_PRIVATE_KEY and SWUPDATE_PASSWORD_FILE should be set
in the local.conf and point to the private RSA key and the (not
mandatory) password file.

Signed-off-by: Jacopo Dall'Aglio <jacopo.dallaglio@kynetics.com>